requirements.txt
edge-tts==6.1.9
gTTS==2.5.3
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
moviepy==1.0.3
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.0
